Made categories and groups orderable by dragging

+++ Spending/Models/CategoryGroup.cs
@@ -13,7 +13,7 @@ namespace Spending.Models
 		public int Id { get; set; }
 		public int UserId { get; set; }
 		public string Name { get; set; }
-		public byte Order { get; set; }
+		public int Order { get; set; }
 
 		public virtual User User { get; set; }
 		public virtual ICollection<Category> Categories { get; set; }
